Changes to NOAA cleaning file and HIndex file
Severely cut the NOAA file to just make sure that the array was from 2014-2018, then began writing algorithm for the merging of the NOAA files and our current data. For the HIndex file, the script was edited to predominantly use tables to work, significant cleaning of needless lines was done. Additionally, the script now just runs off of the Weather_CSV folder, instead of the Weather folder. The Script definitely works for all 23 stations, though som data analysis will be necissary to ensure that the script is producing accurate results.